Auth calls include the parameters for agent's delegated enrolment check.
The PPT reference is supplied in the API call so no longer needs to be extracted from the auth enrolments return; this is fortunate because an agent delegated enrolment is never rendered in the auth.enrolments response.

Protect against normal users supplying another user's ppt reference still projected because we are now supplying a specific enrolment identifier in the Enrolment predicate.